# docs/openapi/squares-api-v1.yaml
openapi: 3.1.1
info:
  title: Squares Server API
  version: 1.0.0
  summary: Ucty, synchronizace vysledku a globalni zebricek hry Squares.
  description: >-
    Prvni verejny kontrakt pro Windows a Android klienta. Prihlaseni zajistuje
    externi OpenID Connect provider; API nikdy neprijima hesla.
  license:
    name: Squares Proprietary Software License
...
servers:
  - url: https://api.squares.example/api/v1
    description: Placeholder produkcni adresy; konkretni domena bude zvolena pri nasazeni.
...
security:
  - bearerAuth: []
paths:
...
  /leaderboards/casual:
    get:
      tags: [Public]
      operationId: getCasualLeaderboard
      summary: Vrati jednu stranku globalniho casual zebricku.
      security: []
      parameters:
        - name: limit
          in: query
          description: Maximalni pocet polozek.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          description: Nepruhledny kurzor z predchozi odpovedi.
          schema:
            type: string
            maxLength: 512
      responses:
        '200':
          description: Stranka zebricku.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /me/leaderboards/casual:
    get:
      tags: [Public]
      operationId: getMyCasualLeaderboardEntry
      summary: Vrati vlastni pozici v casual zebricku.
      responses:
        '200':
          description: Pozice a casual statistiky prihlaseneho hrace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token pro audience Squares API.
...
  responses:
    BadRequest:
      description: Neplatny pozadavek.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Chybejici nebo neplatny access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Operace neni pro ucet povolena.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Zdroj nebyl nalezen.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: Prekrocen limit pozadavku.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    PublicPlayer:
      type: object
      additionalProperties: false
      required: [playerId, handle, displayName, revision, createdAt]
      properties:
        playerId:
          type: string
          format: uuid
        handle:
          type: string
          pattern: '^[a-z0-9][a-z0-9_-]{2,23}$'
        displayName:
          type: string
          minLength: 1
          maxLength: 40
        revision:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
...
    Statistics:
      type: object
      additionalProperties: false
      required: [games, wins, draws, losses, totalScore, winPercentage]
      properties:
        games:
          type: integer
          minimum: 0
        wins:
          type: integer
          minimum: 0
        draws:
          type: integer
          minimum: 0
        losses:
          type: integer
          minimum: 0
        totalScore:
          type: integer
          minimum: 0
        winPercentage:
          type: number
          minimum: 0
          maximum: 100
...
    LeaderboardPage:
      type: object
      additionalProperties: false
      required: [board, rankingMetric, entries, generatedAt]
      properties:
        board:
          type: string
          enum: [casual, ranked]
        rankingMetric:
          type: string
          enum: [TOTAL_SCORE, RATING]
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
        nextCursor:
          type: string
          maxLength: 512
        generatedAt:
          type: string
          format: date-time
    LeaderboardEntry:
      type: object
      additionalProperties: false
      required: [rank, player, statistics]
      properties:
        rank:
          type: integer
          minimum: 1
        player:
          $ref: '#/components/schemas/PublicPlayer'
        statistics:
          $ref: '#/components/schemas/Statistics'
        rating:
          type: integer
          minimum: 0
    Problem:
      type: object
      description: RFC 9457 Problem Details s aplikacnim kodem.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
    Violation:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string